--- pkg/route/allocation/doc.go
@@ -1,2 +1,0 @@
-// Package allocation contains all the route allocation plugins.
-package allocation
