+++ pkg/route/apiserver/routeallocationcontroller/doc.go
@@ -1,2 +1,2 @@
 // Package allocation contains all the route allocation controllers.
-package allocation
+package routeallocationcontroller
