+++ pkg/route/apiserver/simplerouteallocation/doc.go
@@ -1,2 +1,2 @@
 // Package simple contains the SimpleAllocation route plugin.
-package simple
+package simplerouteallocation
